requirements.txt
natsort==8.4.0
numpy==1.26.3
opencv-python==4.9.0.80
pillow==10.2.0
PyScreeze==0.1.30
